requirements.txt
flask
pillow
webp
rembg